% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsFunctions.R
\name{adjustLegendSize}
\alias{adjustLegendSize}
\title{Adjusts the size of a legend}
\usage{
adjustLegendSize(myPlot, pointSize = 0.7, textSize = 7, spaceLegend = 0.7)
}
\arguments{
\item{pointSize}{size of the legend symptoms}

\item{textSize}{size of text}

\item{spaceLegend}{space between legend elements}

\item{myplot}{a ggplot}
}
\description{
Adjusts the size of a legend
}
\examples{
p1 <- adjustLegendSize(p1, pointSize = 0.7, textSize = 10, spaceLegend = 0.7)
}
