% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDemographicsTable.R
\name{CreateDescriptivesTable}
\alias{CreateDescriptivesTable}
\title{Add customized ggplot theme}
\usage{
CreateDescriptivesTable(df, filename = "TableDescriptives.docx")
}
\arguments{
\item{filename}{path and filename}

\item{data_frame}{data frame with a grouping variable called group}
}
\description{
Add customized ggplot theme
}
\examples{
CreateDescriptivesTable(df, "TableDescriptives.docx")

}
