% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotly2SVG.R
\name{ensure_package}
\alias{ensure_package}
\title{Adjusts the size of a legend}
\usage{
ensure_package(package, load = TRUE)
}
\arguments{
\item{p}{a ggplot}

\item{filename}{size of the legend symptoms}

\item{height}{size of text}

\item{width}{space between legend elements}

\item{font_family}{font}
}
\description{
Adjusts the size of a legend
}
\examples{
export_plotly2SVG(p, filename = "violin.sensitive", parent_path = getwd(),
}
