% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreeSurferFunctions.R
\name{load.mgh}
\alias{load.mgh}
\title{Reads individual FreeSurfer .mgh files}
\usage{
load.mgh(input.file)
}
\arguments{
\item{input.file}{input filename}
}
\description{
Reads individual FreeSurfer .mgh files
}
\examples{
fs.data <- load.mgh('./lh.thickness.mgh')

}
