% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsFunctions.R
\name{mytheme.scatter}
\alias{mytheme.scatter}
\title{Add customized ggplot theme}
\usage{
mytheme.scatter(base_size = 11, base_family = "Arial", base_line_size = 0.1)
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}

\item{base_line_size}{line size}
}
\description{
Add customized ggplot theme
}
\examples{
p1 <- p1 + mytheme.scatter(base_size = 11, base_family = "Arial", base_line_size = 0.1)
}
