% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreeSurferFunctions.R
\name{save.mgh}
\alias{save.mgh}
\title{Writes individual FreeSurfer .mgh file}
\usage{
save.mgh(x, fname)
}
\arguments{
\item{input.file}{input filename}
}
\description{
Writes individual FreeSurfer .mgh file
}
\examples{
x <- runif(163842, 1, 10)
save.mgh(x, "lh.test.mgh")

}
